requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ML Inference
tensorflow==2.18.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==11.0.0

# Database & Storage
supabase==2.10.0
httpx==0.28.1

# Security
pydantic==2.10.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
